Rename and move some modues, remove others
