+++ .travis.yml
@@ -2,7 +2,7 @@ sudo: false
 language: node_js
 
 node_js:
-- '5.3'
+- '6.0'
 - '6.9.1'
 - 'stable'
