--- bower.json
@@ -1,21 +1,0 @@
-{
-    "name": "versiony",
-    "main": "index.js",
-    "version": "1.1.1",
-    "homepage": "https://github.com/radubrehar/versiony",
-    "authors": [
-        "Radu.Brehar <radu@jslog.com>"
-    ],
-    "description": "test",
-    "keywords": [
-        "test"
-    ],
-    "license": "MIT",
-    "ignore": [
-        "**/.*",
-        "node_modules",
-        "bower_components",
-        "test",
-        "tests"
-    ]
-}
