--- file2json.js
@@ -1,5 +1,0 @@
-var FS = require('fs')
-
-module.exports = function (file) {
-  return JSON.parse(FS.readFileSync(file))
-}
