--- hasMajorMinorPatch.js
@@ -1,3 +1,0 @@
-module.exports = function (json) {
-  return ('major' in json) && ('minor' in json) && ('patch' in json)
-}
