--- hasVersion.js
@@ -1,3 +1,0 @@
-module.exports = function (json) {
-  return 'version' in json
-}
