+++ lib/index.js
@@ -1,28 +1,15 @@
 'use strict'
 
-var file2json = require('./file2json')
-var json2file = require('./json2file')
-var hasMajorMinorPatch = require('./hasMajorMinorPatch')
-var hasVersion = require('./hasVersion')
-var getVersion = require('./getVersion')
-
-function deprecate (feature, version) {
-  console.warn(`versiony: DEPRECATED: ${feature} is deprecated and will be removed in ${version}`)
-}
+const readJson = require('./read-json')
+const writeJson = require('./write-json')
+const getVersion = require('./get-version')
 
 var versiony = (function () {
   var source, sourceJson
   var version2json = function (model, json) {
     var value = model.get()
 
-    if (hasMajorMinorPatch(json)) {
-      deprecate('the major/minor/patch json interface', '2.0.0')
-      json.major = value[0]
-      json.minor = value[1]
-      json.patch = value[2]
-    } else if (hasVersion(json)) {
-      json.version = value.join('.')
-    }
+    json.version = value.join('.')
 
     return json
   }
@@ -87,7 +74,7 @@ var versiony = (function () {
       source = s || 'package.json'
 
       try {
-        sourceJson = file2json(source)
+        sourceJson = readJson(source)
       } catch (ex) {
         console.log('Could not read source file "' + source + '"! ')
         console.log(ex)
@@ -133,7 +120,7 @@ var versiony = (function () {
       var json
 
       try {
-        json = file2json(file)
+        json = readJson(file)
       } catch (ex) {
         console.log('\nNo file found at "' + file + '". Skipping. \n')
         return this
@@ -143,7 +130,7 @@ var versiony = (function () {
         if (!this.model.hasFile(file)) {
           this.model.file(file)
 
-          json2file(
+          writeJson(
             file,
             version2json(this.model, json),
             indent,
