+++ lib/model.js
@@ -1,5 +1,3 @@
-'use strict'
-
 function join (arr) {
   if (Array.isArray(arr)) {
     arr = arr.join('.')
