+++ lib/read-json.js
@@ -1,0 +1,5 @@
+const fs = require('fs')
+
+module.exports = function (file) {
+  return JSON.parse(fs.readFileSync(file))
+}
