+++ lib/write-json.js
@@ -1,4 +1,4 @@
-var FS = require('fs')
+const fs = require('fs')
 
 module.exports = function (file, contents, indent, eofNewline) {
   var data
@@ -15,7 +15,7 @@ module.exports = function (file, contents, indent, eofNewline) {
     contents += '\n'
   }
 
-  FS.writeFileSync(file, contents)
+  fs.writeFileSync(file, contents)
 
   return data
 }
