+++ tests/bump-spec.js
@@ -4,7 +4,7 @@
  */
 
 'use strict'
-const versiony = require('../index')
+const versiony = require('../lib/index')
 
 // TODO: replace this with destructuring when we support only node>=6 (@job13er 2017-06-14)
 const utils = require('./utils')
